Wav2Lip/requirements.txt
librosa==0.7.0
numpy==1.17.1
opencv-contrib-python>=4.2.0.34
opencv-python==4.5.2.54
torch==1.1.0
torchvision==0.3.0
tqdm==4.45.0
numba==0.48
flask==1.1.2
flask-restful==0.3.8
google-cloud-storage==1.37.1
requests==2.25.1
gunicorn==20.1.0
